+++ client/src/components/HeaderAccountLink.js
@@ -2,20 +2,13 @@ import { Anchor } from 'grommet'
 import Link from 'next/link'
 import * as React from 'react'
 import { useUser } from '../hooks/useUser'
-import { LoginButton } from './LoginButton'
-import { SignInModal } from './modals/SignInModal'
+import { CreateOrLoginModal } from './CreateOrLogin'
 
 export const HeaderAccountLink = () => {
   const { isLoggedIn } = useUser()
-  const [showing, setShowing] = React.useState(false)
 
   if (!isLoggedIn) {
-    return (
-      <>
-        <LoginButton onClick={() => setShowing(true)} />
-        <SignInModal showing={showing} setShowing={setShowing} />
-      </>
-    )
+    return <CreateOrLoginModal title="Create a BioResources Portal Account" />
   }
 
   return (
